requirements.txt
absl-py==2.1.0
dm_env==1.6
fuzzywuzzy==0.18.0
google-generativeai
grpcio-tools
immutabledict==2.0.0
IPython
jsonschema==4.17.3
matplotlib==3.6.1
numpy==1.26.3
opencv-python
pandas==2.1.4
pydub
python-Levenshtein
pytest
requests
tenacity
termcolor
grpcio
portpicker>=1.2.0
protobuf==5.29.0
openai

torch
torchvision
qwen_vl_utils